refactor(mm): M6-S4a neutralize the 3 cross-dir multimodal deps of the in-model VLMs
Prep for relocating the 3 understanding in-model VLM dirs into srt/models/ (S4b): each dir
imported something from multimodal/ outside its own dir, which would re-introduce the
srt->multimodal edge M1 forbids once the dirs move. Neutralize each FIRST (pure moves +
re-export shims, behavior-preserving):

- simple_attention: copied the self-contained pure-jax simple_attention into
  qwen3_omni_moe/audio_encoder.py and dropped the import of
  multimodal.layers.attention.layer (layer.py untouched -- USPAttention still uses it).

- MiMoAudioTransformer: moved the 5 plain-kwargs classes (Qwen2ConfigAdapter, MiMoAudioMLP,
  MiMoAudioAttention, MiMoAudioDecoderLayer, MiMoAudioTransformer) verbatim into the neutral
  srt/layers/mimo_audio_transformer.py. mimo_audio_backbone.py re-imports MiMoAudioTransformer
  (its only remaining user; mimo_audio/__init__ + stage.py re-export still resolve); the
  generation-side MiMoAudioForCausalLM stays. mimo_v2_5/audio_encoder.py now imports the tower
  from the neutral module.

- qwen_vl config: moved multimodal/configs/qwen_vl/qwen_2_5_vl_config.py to
  srt/configs/qwen_vl/ and inlined its 3 base fields (model_path/revision/dtype) so it no
  longer subclasses multimodal's MultiModalModelConfigs (no srt->multimodal edge). Moved the
  qwen-vl vision-config helpers (qwen_vl_vision_config_from_hf + _apply_qwen_vl_vision_overrides
  + key maps) into srt/configs/qwen_vl/config_helpers.py. config_registry.py (stays in
  multimodal for generation) re-imports QwenVLModelVitConfig + _apply_qwen_vl_vision_overrides
  from srt (multimodal->srt is allowed); the in-model qwen2_5_vit.py/qwen2_5_vl.py point srt-side.

- MiMoAudioProcessor (mel front-end, lazy import in mimo_v2_5/audio_codec_processor.py): moved
  to the neutral srt/mm_core/audio_processor.py. multimodal_tokenizer.py (generation) re-imports
  it; the understanding codec processor's lazy import points srt-side.

After this, grep of the 3 dirs for cross-dir multimodal imports returns only own-dir
self-imports (rewritten in S4b). py_compile green for all touched/new files.
